Updated CurveSimplification not to throw exception
Updated CurveSimplification not to throw exception even if the
simplified polygon has less than three vertices. Even in the worst case
where there is no way to get a simplified polygon with more than or
equal to three vertices, we can use the input polygon to keep the
building remain. Otherwise, the building will disappear, which has to be
avoided.